Integrated the functionality of the file watcher into the main app. Now the analysis process starts directly from the main app.
Created a new main file that only needs to be run in order to run the entire program with web portal and api.

Added a launch portal button to the desktop app.
Fixed the text that appears when a recording takes place and put it in the right direction.

The start/stop detection still crashes.
The api not works to upload detections.
The tracker box isn't displayed in the correct position in the analyzed video.