Made the main file output the scripts output to the console. Disabled the logging of the web app and web api to not clutter the main console. Added a new route to resolve the path of videos when accessing them. Modified the event details page to display the right data. Made the main app record in mp4 not avi. Sorted events descending by timestamp. Added back the watcher. Tried to make it queue the processes so the PC doesn't blow up when 2-3 analyses are running at the same time.
Almost working.
Need to first save the detection in a temp file and then move it to the detections folder to be picked up by the watcher since the main app can't launch a subprocess to analyse the movement.
The profile picture might not load correctly. Look into the login route.